+++ tests/gcp/hooks/test_bigquery.py
@@ -30,8 +30,6 @@ from airflow.gcp.hooks.bigquery import (
     _validate_value,
 )
 
-bq_available = True
-
 
 class TestBigQueryHookConnection(unittest.TestCase):
     hook = None  # type: Optional[hook.BigQueryHook]
@@ -71,66 +69,6 @@ class TestPandasGbqCredentials(unittest.TestCase):
         args, kwargs = mock_read_gbq.call_args
         self.assertEqual("CREDENTIALS", kwargs['credentials'])
         self.assertEqual("PROJECT_ID", kwargs['project_id'])
-
-
-class TestBigQueryDataframeResults(unittest.TestCase):
-    def setUp(self):
-        self.instance = hook.BigQueryHook()
-
-    @mock.patch(
-        'airflow.gcp.hooks.base.GoogleCloudBaseHook.project_id',
-        new_callable=mock.PropertyMock,
-        return_value=None
-    )
-    @unittest.skip('SYSTEM TEST, BQ is not available to run tests')
-    def test_output_is_dataframe_with_valid_query(self, mock_project_id):
-        import pandas as pd
-        df = self.instance.get_pandas_df('select 1')
-        self.assertIsInstance(df, pd.DataFrame)
-
-    @mock.patch(
-        'airflow.gcp.hooks.base.GoogleCloudBaseHook.project_id',
-        new_callable=mock.PropertyMock,
-        return_value=None
-    )
-    @unittest.skip('SYSTEM TEST, BQ is not available to run tests')
-    def test_throws_exception_with_invalid_query(self, mock_project_id):
-        with self.assertRaises(Exception) as context:
-            self.instance.get_pandas_df('from `1`')
-        self.assertIn('Reason: ', str(context.exception), "")
-
-    @mock.patch(
-        'airflow.gcp.hooks.base.GoogleCloudBaseHook.project_id',
-        new_callable=mock.PropertyMock,
-        return_value=None
-    )
-    @unittest.skip('SYSTEM TEST, BQ is not available to run tests')
-    def test_succeeds_with_explicit_legacy_query(self, mock_project_id):
-        df = self.instance.get_pandas_df('select 1', dialect='legacy')
-        self.assertEqual(df.iloc(0)[0][0], 1)
-
-    @mock.patch(
-        'airflow.gcp.hooks.base.GoogleCloudBaseHook.project_id',
-        new_callable=mock.PropertyMock,
-        return_value=None
-    )
-    @unittest.skip('SYSTEM TEST, BQ is not available to run tests')
-    def test_succeeds_with_explicit_std_query(self, mock_project_id):
-        df = self.instance.get_pandas_df(
-            'select * except(b) from (select 1 a, 2 b)', dialect='standard')
-        self.assertEqual(df.iloc(0)[0][0], 1)
-
-    @mock.patch(
-        'airflow.gcp.hooks.base.GoogleCloudBaseHook.project_id',
-        new_callable=mock.PropertyMock,
-        return_value=None
-    )
-    @unittest.skip('SYSTEM TEST, BQ is not available to run tests')
-    def test_throws_exception_with_incompatible_syntax(self, mock_project_id):
-        with self.assertRaises(Exception) as context:
-            self.instance.get_pandas_df(
-                'select * except(b) from (select 1 a, 2 b)', dialect='legacy')
-        self.assertIn('Reason: ', str(context.exception), "")
 
 
 class TestBigQueryTableSplitter(unittest.TestCase):
