+++ tests/gcp/hooks/test_bigquery_system.py
@@ -1,0 +1,82 @@
+# -*- coding: utf-8 -*-
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+"""System tests for Google BigQuery hooks"""
+
+import unittest
+from unittest import mock
+
+from airflow.gcp.hooks import bigquery as hook
+from tests.gcp.utils.gcp_authenticator import GCP_BIGQUERY_KEY
+from tests.test_utils.gcp_system_helpers import skip_gcp_system
+
+
+@skip_gcp_system(GCP_BIGQUERY_KEY)
+class BigQueryDataframeResultsSystemTest(unittest.TestCase):
+    def setUp(self):
+        self.instance = hook.BigQueryHook()
+
+    @mock.patch(
+        'airflow.gcp.hooks.base.GoogleCloudBaseHook.project_id',
+        new_callable=mock.PropertyMock,
+        return_value=None
+    )
+    def test_output_is_dataframe_with_valid_query(self, mock_project_id):
+        import pandas as pd
+        df = self.instance.get_pandas_df('select 1')
+        self.assertIsInstance(df, pd.DataFrame)
+
+    @mock.patch(
+        'airflow.gcp.hooks.base.GoogleCloudBaseHook.project_id',
+        new_callable=mock.PropertyMock,
+        return_value=None
+    )
+    def test_throws_exception_with_invalid_query(self, mock_project_id):
+        with self.assertRaises(Exception) as context:
+            self.instance.get_pandas_df('from `1`')
+        self.assertIn('Reason: ', str(context.exception), "")
+
+    @mock.patch(
+        'airflow.gcp.hooks.base.GoogleCloudBaseHook.project_id',
+        new_callable=mock.PropertyMock,
+        return_value=None
+    )
+    def test_succeeds_with_explicit_legacy_query(self, mock_project_id):
+        df = self.instance.get_pandas_df('select 1', dialect='legacy')
+        self.assertEqual(df.iloc(0)[0][0], 1)
+
+    @mock.patch(
+        'airflow.gcp.hooks.base.GoogleCloudBaseHook.project_id',
+        new_callable=mock.PropertyMock,
+        return_value=None
+    )
+    def test_succeeds_with_explicit_std_query(self, mock_project_id):
+        df = self.instance.get_pandas_df(
+            'select * except(b) from (select 1 a, 2 b)', dialect='standard')
+        self.assertEqual(df.iloc(0)[0][0], 1)
+
+    @mock.patch(
+        'airflow.gcp.hooks.base.GoogleCloudBaseHook.project_id',
+        new_callable=mock.PropertyMock,
+        return_value=None
+    )
+    def test_throws_exception_with_incompatible_syntax(self, mock_project_id):
+        with self.assertRaises(Exception) as context:
+            self.instance.get_pandas_df(
+                'select * except(b) from (select 1 a, 2 b)', dialect='legacy')
+        self.assertIn('Reason: ', str(context.exception), "")
